requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
google-generativeai>=0.5.0
python-multipart==0.0.9
maturin==1.5.1